requirements/app_requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=app_requirements.txt app_requirements.in
#
aniso8601==8.0.0
    # via flask-restful
apache-bookkeeper-client==4.12.0
    # via pulsar-client
asbool==0.1.1
    # via -r app_requirements.in
attrs==19.3.0
    # via jsonschema
cachetools==4.2.0
    # via google-auth
certifi==2020.12.5
    # via
    #   pulsar-client
    #   requests
chapar==0.4.6
    # via -r app_requirements.in
chardet==4.0.0
    # via requests
click==7.1.2
    # via flask
contextlib2==21.6.0
    # via schema
decorator==5.1.1
    # via dogpile.cache
deprecated==1.2.13
    # via redis
dogpile.cache==1.1.4
    # via -r app_requirements.in
fastavro==0.24.0
    # via pulsar-client
flask==1.1.2
    # via
    #   -r app_requirements.in
    #   flask-restful
flask-restful==0.3.8
    # via -r app_requirements.in
google-api-core==1.24.1
    # via google-api-python-client
google-api-python-client==1.12.8
    # via -r app_requirements.in
google-auth==1.24.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.0.4
    # via google-api-python-client
googleapis-common-protos==1.52.0
    # via google-api-core
grpcio==1.27.2
    # via
    #   apache-bookkeeper-client
    #   pulsar-client
gunicorn==20.0.4
    # via -r app_requirements.in
httplib2==0.18.1
    # via
    #   google-api-python-client
    #   google-auth-httplib2
    #   oauth2client
idna==2.10
    # via requests
itsdangerous==1.1.0
    # via flask
jinja2==2.11.2
    # via flask
joblib==0.16.0
    # via scikit-learn
jsonschema==3.2.0
    # via -r app_requirements.in
markupsafe==1.1.1
    # via jinja2
numpy==1.19.1
    # via
    #   -r app_requirements.in
    #   pandas
    #   scikit-learn
    #   scipy
oauth2client==4.1.3
    # via -r app_requirements.in
packaging==21.3
    # via redis
pandas==1.1.0
    # via -r app_requirements.in
pbr==5.8.0
    # via stevedore
prometheus-client==0.9.0
    # via pulsar-client
protobuf==3.14.0
    # via
    #   apache-bookkeeper-client
    #   google-api-core
    #   googleapis-common-protos
    #   pulsar-client
psycopg2-binary==2.8.5
    # via -r app_requirements.in
pulsar-client==2.7.0
    # via chapar
pyasn1==0.4.8
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via
    #   google-auth
    #   oauth2client
pymmh3==0.0.5
    # via apache-bookkeeper-client
pyparsing==3.0.6
    # via packaging
pyrsistent==0.16.0
    # via jsonschema
python-dateutil==2.8.1
    # via pandas
pytz==2020.1
    # via
    #   apache-bookkeeper-client
    #   fastavro
    #   flask-restful
    #   google-api-core
    #   pandas
ratelimit==2.2.1
    # via pulsar-client
redis==4.1.0
    # via -r app_requirements.in
requests==2.25.1
    # via
    #   -r app_requirements.in
    #   apache-bookkeeper-client
    #   google-api-core
rsa==4.6
    # via
    #   google-auth
    #   oauth2client
schema==0.7.5
    # via -r app_requirements.in
scikit-learn==0.23.2
    # via -r app_requirements.in
scipy==1.5.2
    # via scikit-learn
six==1.15.0
    # via
    #   apache-bookkeeper-client
    #   asbool
    #   flask-restful
    #   google-api-core
    #   google-api-python-client
    #   google-auth
    #   google-auth-httplib2
    #   grpcio
    #   jsonschema
    #   oauth2client
    #   protobuf
    #   pulsar-client
    #   pyrsistent
    #   python-dateutil
sqlalchemy==1.3.18
    # via -r app_requirements.in
stevedore==3.5.0
    # via dogpile.cache
threadpoolctl==2.1.0
    # via scikit-learn
uritemplate==3.0.1
    # via google-api-python-client
urllib3==1.26.2
    # via requests
werkzeug==1.0.1
    # via flask
wrapt==1.13.3
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
